tests(create logistics order): create test class for the command
