+++ tests/Unit/WooCommerceOrderStatusChangedTest.php
@@ -97,35 +97,4 @@ class WooCommerceOrderStatusChangedTest extends TestCase
 
         $this->writeApi->shouldHaveReceived("logisticsOrderPost", [LogisticsOrder::class]);
     }
-
-    public function test_it_updates_the_order_state_after_a_successful_api_request()
-    {
-        /** @var MockInterface|stdClass */
-        $order = Mockery::mock("WC_Order");
-        $order->allows("set_status");
-        $order->allows("save");
-
-        $this->plugin->createLogisticsOrderAndUpdateOrderState($order);
-
-        $order->shouldHaveReceived("save");
-    }
-
-    public function test_it_does_not_update_the_order_state_after_a_failed_request()
-    {
-        /** @var MockInterface|WriteApi */
-        $writeApi = Mockery::mock(WriteApi::class);
-        $writeApi->shouldReceive("logisticsOrderPost")->andThrow(new ApiException("Unauthenticated", 401));
-        $writeApi->allows(["getConfig" => new Configuration()]);
-
-        $this->plugin->setWriteApiClient($writeApi);
-
-        /** @var MockInterface|stdClass */
-        $order = Mockery::mock("WC_Order");
-        $order->allows("set_status");
-        $order->allows("save");
-
-        $this->plugin->createLogisticsOrderAndUpdateOrderState($order);
-
-        $order->shouldNotHaveReceived("save");
-    }
 }
